pyproject.toml
[tool.poetry]
name = "successful-banking-operations"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
mypy = "^1.12.1"
pytest = "^8.3.3"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.12.1"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"
